Update and rename realtime parsing tests and add some more tests
